some major modifications
removed unwanted read / write locks to datastructures. added atomic
integers to count persons entering and exiting train. added method to
check person thread interruption. Now we are not interrupting persons in
train who are trying to exit. we wait for all persons to exit before we
announce for incoming passengers in platform.